Basic student registration program built with Java, Swing and MySQL
ITU Student Registration is the program that can be logged in with two different types of users. It is written to be used by students and lecturers.
Different utilizations are defined for each type of the users. While students are able to view contact information and exam results of all students, lecturers are able to insert new student records and grades for each student.
MySQL, Java and Swing are used to create this program.